优化Application.py结构

+++ src/Application.py
@@ -7,9 +7,28 @@ import multiprocessing
 import platform
 import signal
 
+
+# Application类
+class Application:
+
+    def __init__(self, port, baud_rate, ATListFileNames, loopTimes):
+        self.port = port
+        self.baud_rate = baud_rate
+        self.ATListFileNames = ATListFileNames
+        self.loopTimes = loopTimes
+
+    def run(self):
+        app = ATestUtils.ATestUtils(self.port, self.baud_rate)
+        app.ATest(self.ATListFileNames, self.loopTimes)
+        print(f'共发生错误{app.error_count}次')
+
+
 # 定义ATListFiles
 ATListFileNames = []
+# 初始化波特率设置
+baud_rate = 115200
 loopTimes = int(input('请输入循环次数：'))
+# 查询系统平台
 system_cate = platform.system()
 print(f'当前操作系统为：{system_cate}')
 if system_cate == 'Linux':
@@ -39,70 +58,54 @@ try:
             continue
         ATListFileNames.append(f'{fileName}.txt')
 
+    # 当前如果是Linux的话，询问用户是否开启trace抓取功能
     enable_trace = 'n'
     if system_cate == 'Linux':
         enable_trace = input('当前操作系统为Linux，是否抓取trace？（y/n）')
         if enable_trace == 'y':
             diag_port = input('请输入diag诊断口端口：')
-except KeyboardInterrupt:
-    print()
-    print('Exit...')
-    sys.exit()
-baud_rate = 115200
+
+    print(
+        "JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST")
+    print(
+        "JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST")
+    print(
+        "JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST")
+
+    diag_pid = 0
 
 
-class Application:
-
-    def __init__(self, port, baud_rate, ATListFileNames, loopTimes):
-        self.port = port
-        self.baud_rate = baud_rate
-        self.ATListFileNames = ATListFileNames
-        self.loopTimes = loopTimes
-
-    def run(self):
-        app = ATestUtils.ATestUtils(self.port, self.baud_rate)
-        app.ATest(self.ATListFileNames, self.loopTimes)
-        print(f'共发生错误{app.error_count}次')
+    def start_trace():
+        global diag_pid
+        diag_pid = os.getpid()
+        print(f'Run diag process {diag_pid}')
+        os.popen(f"./bin/diag trace/log - - {diag_port}")
 
 
-print("JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST")
-print("JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST")
-print("JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST---JEREMYPYATEST")
-
-diag_pid = 0
-
-
-def start_trace():
-    global diag_pid
-    diag_pid = os.getpid()
-    print(f'Run diag process {diag_pid}')
-    os.popen(f"./bin/diag trace/log - - {diag_port}")
-
-
-try:
     print(f'Application process {os.getpid()}')
     if system_cate == 'Linux' and enable_trace == 'y':
         multiprocessing.Process(target=start_trace).start()
-    Application(port, baud_rate, ATListFileNames, loopTimes).run()
+    try:
+        Application(port, baud_rate, ATListFileNames, loopTimes).run()
+    except serial.serialutil.SerialException as se:
+        print(se)
+        print("---------------")
+        print(traceback.format_exc())
+        if system_cate == 'Linux' and enable_trace == 'y':
+            os.kill(diag_pid, signal.SIGKILL)
+        if 'No such file or directory' in traceback.format_exc():
+            print('输入的端口不存在')
+        if 'PermissionError' in traceback.format_exc():
+            print('端口被占用,请检查是否有其他程序正在占用设备端口')
+    except Exception as e:
+        if system_cate == 'Linux' and enable_trace == 'y':
+            os.kill(diag_pid, signal.SIGKILL)
+        print(e)
+        print("---------------")
+        print(traceback.format_exc())
 except KeyboardInterrupt as ke:
     print()
     print("Exit...")
-    if system_cate == 'Linux' and enable_trace == 'y':
+    if system_cate == 'Linux' and enable_trace == 'y' and diag_pid != 0:
         os.kill(diag_pid, signal.SIGKILL)
     sys.exit()
-except serial.serialutil.SerialException as se:
-    print(se)
-    print("---------------")
-    print(traceback.format_exc())
-    if system_cate == 'Linux' and enable_trace == 'y':
-        os.kill(diag_pid, signal.SIGKILL)
-    if 'No such file or directory' in traceback.format_exc():
-        print('输入的端口不存在')
-    if 'PermissionError' in traceback.format_exc():
-        print('端口被占用,请检查是否有其他程序正在占用设备端口')
-except Exception as e:
-    if system_cate == 'Linux' and enable_trace == 'y':
-        os.kill(diag_pid, signal.SIGKILL)
-    print(e)
-    print("---------------")
-    print(traceback.format_exc())
